+++ src/save_gui_manager.cpp
@@ -30,7 +30,7 @@ void
 SaveGUIManager::draw_overlay ()
 {
   g_graphic_context->draw_string(10.0f,
-                                 g_graphic_context->get_height() - 10.0f,
+                                 static_cast<float>(g_graphic_context->get_height()) - 10.0f,
                                  "Save Dialog");
 }
